refactor: adding go modules
