+++ cmd/root.go
@@ -2,7 +2,7 @@ package cmd
 
 import (
 	"github.com/getsentry/sentry-go"
-	"github.com/qovery/qovery-cli/io"
+	"github.com/qovery/qovery-cli/pkg"
 	"github.com/qovery/qovery-cli/utils"
 	"github.com/spf13/cobra"
 	"os"
@@ -37,11 +37,11 @@ func initConfig() {
 }
 
 func initSentry() {
-	io.GetCurrentVersion()
+	pkg.GetCurrentVersion()
 	err := sentry.Init(sentry.ClientOptions{
 		Dsn:         "https://199e1e8385d94377a98676dadcd77e2d@o471935.ingest.sentry.io/5866472",
 		Environment: "prod",
-		Release:     io.GetCurrentVersion(),
+		Release:     pkg.GetCurrentVersion(),
 		// Enable printing of SDK debug messages.
 		// Useful when getting started or trying to figure something out.
 		Debug: false,
