+++ pkg/deploy.go
@@ -1,4 +1,4 @@
-package io
+package pkg
 
 import (
 	"bytes"
