+++ pkg/version.go
@@ -1,4 +1,4 @@
-package io
+package pkg
 
 import (
 	"errors"
