+++ utils/go.mod
@@ -1,1 +1,3 @@
-module "utils"
+module github.com/qovery/qovery-cli/utils
+
+go 1.16
